# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SL Real Estate & Construction Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type PropertyListingType =
  (typeof PropertyListingType)[keyof typeof PropertyListingType];

export const PropertyListingType = {
  sale: "sale",
  rent: "rent",
} as const;

export type PropertyPropertyType =
  (typeof PropertyPropertyType)[keyof typeof PropertyPropertyType];

export const PropertyPropertyType = {
  house: "house",
  land: "land",
  apartment: "apartment",
  commercial: "commercial",
} as const;

export type PropertyStatus =
  (typeof PropertyStatus)[keyof typeof PropertyStatus];

export const PropertyStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Property {
  id: number;
  title: string;
  description: string;
  listingType: PropertyListingType;
  propertyType: PropertyPropertyType;
  district: string;
  latitude?: number | null;
  longitude?: number | null;
  priceNpr: number;
  areaDhur?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  buildYear?: number | null;
  amenities: string[];
  photos: string[];
  videoUrl?: string | null;
  status: PropertyStatus;
  featured: boolean;
  ownerName: string;
  ownerPhone: string;
  ownerWhatsapp?: string | null;
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreatePropertyInputListingType =
  (typeof CreatePropertyInputListingType)[keyof typeof CreatePropertyInputListingType];

export const CreatePropertyInputListingType = {
  sale: "sale",
  rent: "rent",
} as const;

export type CreatePropertyInputPropertyType =
  (typeof CreatePropertyInputPropertyType)[keyof typeof CreatePropertyInputPropertyType];

export const CreatePropertyInputPropertyType = {
  house: "house",
  land: "land",
  apartment: "apartment",
  commercial: "commercial",
} as const;

export interface CreatePropertyInput {
  title: string;
  description: string;
  listingType: CreatePropertyInputListingType;
  propertyType: CreatePropertyInputPropertyType;
  district: string;
  latitude?: number | null;
  longitude?: number | null;
  priceNpr: number;
  areaDhur?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  buildYear?: number | null;
  amenities?: string[];
  photos?: string[];
  videoUrl?: string | null;
  ownerName: string;
  ownerPhone: string;
  ownerWhatsapp?: string | null;
}

export type UpdatePropertyInputListingType =
  (typeof UpdatePropertyInputListingType)[keyof typeof UpdatePropertyInputListingType];

export const UpdatePropertyInputListingType = {
  sale: "sale",
  rent: "rent",
} as const;

export type UpdatePropertyInputPropertyType =
  (typeof UpdatePropertyInputPropertyType)[keyof typeof UpdatePropertyInputPropertyType];

export const UpdatePropertyInputPropertyType = {
  house: "house",
  land: "land",
  apartment: "apartment",
  commercial: "commercial",
} as const;

export type UpdatePropertyInputStatus =
  (typeof UpdatePropertyInputStatus)[keyof typeof UpdatePropertyInputStatus];

export const UpdatePropertyInputStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdatePropertyInput {
  title?: string;
  description?: string;
  listingType?: UpdatePropertyInputListingType;
  propertyType?: UpdatePropertyInputPropertyType;
  district?: string;
  latitude?: number | null;
  longitude?: number | null;
  priceNpr?: number;
  areaDhur?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  buildYear?: number | null;
  amenities?: string[];
  photos?: string[];
  videoUrl?: string | null;
  ownerName?: string;
  ownerPhone?: string;
  ownerWhatsapp?: string | null;
  status?: UpdatePropertyInputStatus;
  featured?: boolean;
}

export interface ConstructionService {
  id: number;
  businessName: string;
  serviceType: string;
  phone: string;
  whatsapp?: string | null;
  district: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateServiceInput {
  businessName: string;
  serviceType: string;
  phone: string;
  whatsapp?: string | null;
  district: string;
  description?: string | null;
}

export type AboutContentSocialLinks = {
  facebook?: string | null;
  instagram?: string | null;
  twitter?: string | null;
};

export interface AboutContent {
  mission: string;
  vision: string;
  contactPhone: string;
  contactEmail: string;
  address: string;
  socialLinks?: AboutContentSocialLinks;
  updatedAt?: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  phone?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type GetPropertiesParams = {
  district?: string;
  listingType?: GetPropertiesListingType;
  propertyType?: GetPropertiesPropertyType;
  minPrice?: number;
  maxPrice?: number;
  minArea?: number;
  maxArea?: number;
  bedrooms?: number;
  amenities?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type GetPropertiesListingType =
  (typeof GetPropertiesListingType)[keyof typeof GetPropertiesListingType];

export const GetPropertiesListingType = {
  sale: "sale",
  rent: "rent",
} as const;

export type GetPropertiesPropertyType =
  (typeof GetPropertiesPropertyType)[keyof typeof GetPropertiesPropertyType];

export const GetPropertiesPropertyType = {
  house: "house",
  land: "land",
  apartment: "apartment",
  commercial: "commercial",
} as const;

export type GetProperties200 = {
  properties: Property[];
  total: number;
};

export type GetFeaturedProperties200 = {
  properties: Property[];
};

export type AdminGetPropertiesParams = {
  status?: AdminGetPropertiesStatus;
  limit?: number;
  offset?: number;
};

export type AdminGetPropertiesStatus =
  (typeof AdminGetPropertiesStatus)[keyof typeof AdminGetPropertiesStatus];

export const AdminGetPropertiesStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type AdminGetProperties200 = {
  properties: Property[];
  total: number;
};

export type RejectPropertyBody = {
  reason?: string;
};

export type FeaturePropertyBody = {
  featured: boolean;
};

export type GetServicesParams = {
  district?: string;
  serviceType?: string;
};

export type GetServices200 = {
  services: ConstructionService[];
};

export type GetFavoritesParams = {
  userId: string;
};

export type GetFavorites200 = {
  favoriteIds: number[];
  properties: Property[];
};

export type AddFavoriteBody = {
  userId: string;
  propertyId: number;
};

export type AddFavorite201 = {
  success: boolean;
};

export type RemoveFavoriteParams = {
  userId: string;
};