# lib/api-zod/src/generated/types/createPropertyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SL Real Estate & Construction Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePropertyInputListingType } from "./createPropertyInputListingType";
import type { CreatePropertyInputPropertyType } from "./createPropertyInputPropertyType";

export interface CreatePropertyInput {
  title: string;
  description: string;
  listingType: CreatePropertyInputListingType;
  propertyType: CreatePropertyInputPropertyType;
  district: string;
  latitude?: number | null;
  longitude?: number | null;
  priceNpr: number;
  areaDhur?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  buildYear?: number | null;
  amenities?: string[];
  photos?: string[];
  videoUrl?: string | null;
  ownerName: string;
  ownerPhone: string;
  ownerWhatsapp?: string | null;
}