# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SL Real Estate & Construction Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyListingType } from "./propertyListingType";
import type { PropertyPropertyType } from "./propertyPropertyType";
import type { PropertyStatus } from "./propertyStatus";

export interface Property {
  id: number;
  title: string;
  description: string;
  listingType: PropertyListingType;
  propertyType: PropertyPropertyType;
  district: string;
  latitude?: number | null;
  longitude?: number | null;
  priceNpr: number;
  areaDhur?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  buildYear?: number | null;
  amenities: string[];
  photos: string[];
  videoUrl?: string | null;
  status: PropertyStatus;
  featured: boolean;
  ownerName: string;
  ownerPhone: string;
  ownerWhatsapp?: string | null;
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
}